src/api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0
prometheus-client==0.19.0
